Add BodyModal component
